Add a test for static const declarations in structure types
The functionality already appears to work, and this test is just to make sure we don't regress on it.

The most interesting thing here is that I'm using this change to pitch a new organization for tests around what part of the language they cover (rather than the kind of test they are), since the `tests/compute/` directory is getting overly full and is hard to navigate. We can consider moving individual tests into more of a hierarchy at some later point.