Fix crashing bug when using overloaded name as generic arg (#1315)
If somebody defines two `struct` types with the same name:

```hlsl
struct A {}

// ...

struct A {}
```

and then tries to use that name when specializing a generic function:

```hlsl
void doThing<T>() { ... }

// ...

doThing<A>();
```

then the Slang front-end currently crashes, which leads to it not diagnosing the original problem (the conflicting declarations of `A`).

This change fixes up the checking of generic arguments so that it properly fills in dummy "error" arguments in place of missing or incorrect arguments, and thus guarantees that the generic substitution it creates will at least be usable for the next steps of checking (rather than leaving null pointers in the substitution).

This change also fixes up the error message for the case where a generic application like `F<A>` is formed where `F` is not a generic. We already had a more refined diagnostic defined for that case, but for some reason the site in the code where we ought to use it was still issuing an internal compiler error around an unimplemented feature.

This chagne includes a diagnostic test case to cover both of the above fixes.